afxdp/cos/queue_ops/mod: relocate inline tests to tests.rs
queue_ops/ is a directory module (mod.rs + drain.rs + pop.rs + push.rs + v_min.rs). 1,139-LOC inline test block at end of mod.rs extracted to sibling tests.rs (1,142 LOC incl 5-line header). Stub: `#[cfg(test)] #[path = "tests.rs"] mod tests;`.

After: queue_ops/mod.rs drops from 1,268 → 132 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
